+++ src/authority/persistence.rs
@@ -7,14 +7,13 @@
 use std::iter::Iterator;
 
 use time;
-use time::Timespec;
 use rusqlite;
 use rusqlite::Connection;
 use rusqlite::SqliteError;
 
 use ::error::PersistenceError;
 use ::error::PersistenceResult;
-use ::op::Message;
+use ::rr::Record;
 use ::serialize::binary::{BinDecoder, BinEncoder, BinSerializable};
 
 const CURRENT_VERSION: i64 = 1;
@@ -36,37 +35,38 @@ impl Journal {
     self.version
   }
 
+  /// this returns an iterator from the beginning of time, to be used to recreate an authority
   pub fn iter<'j>(&'j self) -> JournalIter<'j> {
     JournalIter::new(self)
   }
 
-  /// inserts a message, this is an append only operation.
+  /// Inserts a record, this is an append only operation.
   ///
-  /// Messages should never be posthumously modified. The message will be serialized into the.
+  /// Records should never be posthumously modified. The message will be serialized into the.
   ///  the first message serialized to the journal, should be a single AXFR of the entire zone,
   ///  this will be used as a starting point to reconstruct the zone.
   ///
   /// # Argument
   ///
-  /// * `message` - will be serialized into the journal
-  pub fn insert_message(&self, soa_serial: u32, message: &Message) -> PersistenceResult<()> {
+  /// * `record` - will be serialized into the journal
+  pub fn insert_record(&self, soa_serial: u32, record: &Record) -> PersistenceResult<()> {
     assert!(self.version == CURRENT_VERSION, "schema version mismatch, schema_up() resolves this");
 
-    let mut serial_message: Vec<u8> = Vec::with_capacity(512);
+    let mut serial_record: Vec<u8> = Vec::with_capacity(512);
     {
-      let mut encoder = BinEncoder::new(&mut serial_message);
-      try_rethrow!(PersistenceError::EncodeError, message.emit(&mut encoder));
+      let mut encoder = BinEncoder::new(&mut serial_record);
+      try_rethrow!(PersistenceError::EncodeError, record.emit(&mut encoder));
     }
 
     let timestamp = time::get_time();
-    let client_id: i64 = message.get_id() as i64;
+    let client_id: i64 = 0; // TODO: we need better id information about the client, like pub_key
     let soa_serial: i64 = soa_serial as i64;
 
     let count = try_rethrow!(PersistenceError::SqliteError,
                               self.conn.execute("INSERT
-                                                 INTO messages (client_id, soa_serial, timestamp, message)
+                                                 INTO records (client_id, soa_serial, timestamp, record)
                                                  VALUES ($1, $2, $3, $4)",
-                                                 &[&client_id, &soa_serial, &timestamp, &serial_message]));
+                                                 &[&client_id, &soa_serial, &timestamp, &serial_record]));
     //
     if count != 1 {
       return Err(PersistenceError::WrongInsertCount{loc: error_loc!(), got: count, expect: 1});
@@ -75,7 +75,17 @@ impl Journal {
     Ok(())
   }
 
-  /// Selects a message from the given row_id.
+  /// Inserts a set of records into the Journal, a convenience method for insert_record
+  pub fn insert_records(&self, soa_serial: u32, records: &[Record]) -> PersistenceResult<()> {
+    // TODO: NEED TRANSACTION HERE
+    for record in records {
+      try!(self.insert_record(soa_serial, record));
+    }
+
+    Ok(())
+  }
+
+  /// Selects a record from the given row_id.
   ///
   /// This allows for the entire set of records to be iterated through, by starting at 0, and
   ///  incrementing each subsequent row.
@@ -84,33 +94,32 @@ impl Journal {
   ///
   /// * `row_id` - the row_id can either be exact, or start at 0 to get the earliest row in the
   ///              list.
-  pub fn select_message(&self, row_id: i64) -> PersistenceResult<Option<(i64, Message)>> {
+  pub fn select_record(&self, row_id: i64) -> PersistenceResult<Option<(i64, Record)>> {
     assert!(self.version == CURRENT_VERSION, "schema version mismatch, schema_up() resolves this");
 
     let mut stmt = try_rethrow!(PersistenceError::SqliteError,
-                                self.conn.prepare("SELECT _rowid_, message
-                                                FROM messages
+                                self.conn.prepare("SELECT _rowid_, record
+                                                FROM records
                                                 WHERE _rowid_ >= $1
                                                 LIMIT 1"));
 
-    let message_opt: Option<Result<(i64, Message), SqliteError>> = try_rethrow!(PersistenceError::SqliteError,
+    let record_opt: Option<Result<(i64, Record), SqliteError>> = try_rethrow!(PersistenceError::SqliteError,
                                                             stmt.query_and_then(&[&row_id],
-                                                            |row| -> Result<(i64, Message), SqliteError> {
+                                                            |row| -> Result<(i64, Record), SqliteError> {
       let row_id: i64 = try!(row.get_checked(0));
-      let message_bytes: Vec<u8> = try!(row.get_checked(1));
-      let mut decoder = BinDecoder::new(&message_bytes);
-
+      let record_bytes: Vec<u8> = try!(row.get_checked(1));
+      let mut decoder = BinDecoder::new(&record_bytes);
 
       // todo add location to this...
-      match Message::read(&mut decoder) {
-        Ok(message) => Ok((row_id, message)),
+      match Record::read(&mut decoder) {
+        Ok(record) => Ok((row_id, record)),
         Err(decode_error) => Err(rusqlite::Error::FromSqlConversionFailure(Box::new(decode_error))),
       }
     })).next();
 
     //
-    match message_opt {
-      Some(Ok((row_id, message))) => Ok(Some((row_id, message))),
+    match record_opt {
+      Some(Ok((row_id, record))) => Ok(Some((row_id, record))),
       Some(Err(err)) => Err(PersistenceError::SqliteError(error_loc!(), err)),
       None => Ok(None),
     }
@@ -170,7 +179,7 @@ impl Journal {
     while self.version < CURRENT_VERSION {
       match self.version + 1 {
         0 => self.version = try!(self.init_up()),
-        1 => self.version = try!(self.messages_up()),
+        1 => self.version = try!(self.records_up()),
         _ => panic!("incorrect version somewhere"),
       }
 
@@ -198,17 +207,17 @@ impl Journal {
     Ok(0)
   }
 
-  /// adds the messages table, this is the main and single table for the history of changes to an
-  ///  authority.
-  fn messages_up(&self) -> PersistenceResult<i64> {
+  /// adds the records table, this is the main and single table for the history of changes to an
+  ///  authority. Each record is expected to be in the format of an update record
+  fn records_up(&self) -> PersistenceResult<i64> {
     // we'll be using rowid for our primary key, basically: `rowid INTEGER PRIMARY KEY ASC`
     let count = try_rethrow!(PersistenceError::SqliteError,
-                              self.conn.execute("CREATE TABLE messages (
-                              client_id      INTEGER NOT NULL,
-                              soa_serial      INTEGER NOT NULL,
-                              timestamp       TEXT NOT NULL,
-                              message         BLOB NOT NULL
-                            )", &[]));
+                              self.conn.execute("CREATE TABLE records (
+                                                  client_id      INTEGER NOT NULL,
+                                                  soa_serial     INTEGER NOT NULL,
+                                                  timestamp      TEXT NOT NULL,
+                                                  record         BLOB NOT NULL
+                                                )", &[]));
     //
     assert_eq!(count, 1);
 
@@ -228,15 +237,15 @@ impl<'j> JournalIter<'j> {
 }
 
 impl<'j> Iterator for JournalIter<'j> {
-  type Item = Message;
+  type Item = Record;
 
   fn next(&mut self) -> Option<Self::Item> {
-    let next: PersistenceResult<Option<(i64, Message)>> = self.journal.select_message(self.current_row_id + 1);
+    let next: PersistenceResult<Option<(i64, Record)>> = self.journal.select_record(self.current_row_id + 1);
 
     match next {
-      Ok(Some((row_id, message))) => {
+      Ok(Some((row_id, record))) => {
         self.current_row_id = row_id;
-        Some(message)
+        Some(record)
       },
       Ok(None) => {
         None
@@ -265,38 +274,18 @@ fn test_init_journal() {
 }
 
 #[cfg(test)]
-fn create_test_journal() -> (Message, Journal) {
+fn create_test_journal() -> (Record, Journal) {
   use std::net::Ipv4Addr;
   use std::str::FromStr;
 
-  use ::op::{Message, MessageType, OpCode, Query, UpdateMessage};
-  use ::rr::{DNSClass, Name, RData, Record, RecordType};
+  use ::rr::{Name, RData, Record, RecordType};
 
-  let zone_origin = Name::with_labels(vec!["example".to_string(), "com".to_string()]);
   let www = Name::with_labels(vec!["www".to_string(),"example".to_string(), "com".to_string()]);
-
-  let mut message: Message = Message::new();
-  message.id(10)
-         .message_type(MessageType::Query)
-         .op_code(OpCode::Update)
-         .authoritative(false)
-         .truncated(false)
-         .recursion_desired(true)
-         .recursion_available(false);
-  //
-
-  let mut zone: Query = Query::new();
-  zone.name(zone_origin).query_class(DNSClass::IN).query_type(RecordType::SOA);
-
-  // build the message
-  message.add_zone(zone);
 
   let mut record = Record::new();
   record.name(www);
   record.rr_type(RecordType::A);
   record.rdata(RData::A(Ipv4Addr::from_str("127.0.0.1").unwrap()));
-
-  message.add_update(record.clone());
 
   // test that this message can be inserted
   let conn = Connection::open_in_memory().expect("could not create in memory DB");
@@ -304,52 +293,51 @@ fn create_test_journal() -> (Message, Journal) {
   journal.schema_up().unwrap();
 
   // insert the message
-  journal.insert_message(0, &message).unwrap();
-
-  // serialize the message to get a cononical message, then compare...
-  let mut serial_message: Vec<u8> = Vec::with_capacity(512);
-  {
-    let mut encoder = BinEncoder::new(&mut serial_message);
-    message.emit(&mut encoder).unwrap();
-  }
-  let mut decoder = BinDecoder::new(&serial_message);
-  let mut message = Message::read(&mut decoder).unwrap();
+  journal.insert_record(0, &record).unwrap();
 
   // insert another...
-  message.id(11);
-  journal.insert_message(0, &message).unwrap();
+  record.rdata(RData::A(Ipv4Addr::from_str("127.0.1.1").unwrap()));
+  journal.insert_record(0, &record).unwrap();
 
-  (message, journal)
+  (record, journal)
 }
 
 #[test]
-fn test_insert_and_select_message() {
-  let (mut message, journal) = create_test_journal();
+fn test_insert_and_select_record() {
+  use std::net::Ipv4Addr;
+  use std::str::FromStr;
 
-  message.id(10);
+  use rr::RData;
 
-  // select the message
-  let (row_id, journal_message) = journal.select_message(0).expect("persistence error").expect("none");
-  assert_eq!(journal_message, message);
+  let (mut record, journal) = create_test_journal();
 
-  // insert another...
-  message.id(11);
-  let (row_id, journal_message) = journal.select_message(row_id + 1).expect("persistence error").expect("none");
-  assert_eq!(journal_message, message);
-  assert_eq!(message.get_id(), 11);
+  // select the record
+  let (row_id, journal_record) = journal.select_record(0).expect("persistence error").expect("none");
+  record.rdata(RData::A(Ipv4Addr::from_str("127.0.0.1").unwrap()));
+  assert_eq!(journal_record, record);
+
+  // test another
+  let (row_id, journal_record) = journal.select_record(row_id + 1).expect("persistence error").expect("none");
+  record.rdata(RData::A(Ipv4Addr::from_str("127.0.1.1").unwrap()));
+  assert_eq!(journal_record, record);
 
   // check that we get nothing for id over row_id
-  let option_none = journal.select_message(row_id + 1).expect("persistence error");
+  let option_none = journal.select_record(row_id + 1).expect("persistence error");
   assert!(option_none.is_none());
 }
 
 #[test]
 fn test_iterator() {
-  let (mut message, journal) = create_test_journal();
+  use std::net::Ipv4Addr;
+  use std::str::FromStr;
+
+  use rr::RData;
+
+  let (mut record, journal) = create_test_journal();
 
   let mut iter = journal.iter();
 
-  assert_eq!(message.id(10), &iter.next().unwrap());
-  assert_eq!(message.id(11), &iter.next().unwrap());
+  assert_eq!(record.rdata(RData::A(Ipv4Addr::from_str("127.0.0.1").unwrap())), &iter.next().unwrap());
+  assert_eq!(record.rdata(RData::A(Ipv4Addr::from_str("127.0.1.1").unwrap())), &iter.next().unwrap());
   assert_eq!(None, iter.next());
 }