+++ src/serialize/binary/decoder.rs
@@ -73,14 +73,19 @@ impl<'a> BinDecoder<'a> {
   /// length (including the length octet).
   ///
   /// the vector should be reversed before calling.
-  pub fn read_character_data(&mut self) -> DecodeResult<String> {
+  pub fn read_character_data(&mut self, lowercase: bool) -> DecodeResult<String> {
     let length: u8 = try!(self.pop());
 
     // TODO once Drain stabalizes on Vec, this should be replaced...
     let label_vec: Vec<u8> = try!(self.read_vec(length as usize));
 
     // translate bytes to string, then lowercase...
-    Ok(try!(String::from_utf8(label_vec)).to_lowercase())
+    let data = try!(String::from_utf8(label_vec));
+    if lowercase {
+      Ok(data.to_lowercase())
+    } else {
+      Ok(data)
+    }
   }
 
   pub fn read_vec(&mut self, len: usize) -> DecodeResult<Vec<u8>> {